FilterContainer with useReducer
